+++ benches/float/round.rs
@@ -114,7 +114,7 @@ fn round(float: f64, decimals: u8) -> f64 {
 pub async fn load(decimals: u8) -> Result<()> {
     let db = &DB_POOL.get().await.unwrap();
     let mut stats = Vec::new();
-    for group in QueryStats::load(db, decimals).await? {
+    for group in CompressedQueryStats::load(db, decimals).await? {
         for stat in group.decompress() {
             stats.push(stat);
         }
@@ -135,7 +135,7 @@ pub async fn load(decimals: u8) -> Result<()> {
         pub shared_blks_hit: i64,
         pub shared_blks_read: i64,
     }
-    pub struct QueryStats {
+    pub struct CompressedQueryStats {
         database_id: i64,
         start_at: SystemTime,
         collected_at: Vec<u8>,
@@ -149,7 +149,7 @@ pub async fn load(decimals: u8) -> Result<()> {
         shared_blks_hit: Vec<u8>,
         shared_blks_read: Vec<u8>,
     }
-    impl QueryStats {
+    impl CompressedQueryStats {
         pub async fn load(db: &Client, decimals: u8) -> Result<Vec<Self>> {
             let sql = &format!("
                 SELECT database_id, start_at, collected_at, collected_secs, fingerprint, postgres_role_id, calls, rows, total_time, io_time, shared_blks_hit, shared_blks_read
@@ -157,7 +157,7 @@ pub async fn load(decimals: u8) -> Result<()> {
             ");
             let mut results = Vec::new();
             for row in db.query(&db.prepare_cached(sql).await?, &[]).await? {
-                results.push(QueryStats {
+                results.push(Self {
                     database_id: row.get(0),
                     start_at: row.get(1),
                     collected_at: row.get(2),